[do not merge] fix input/output order issue
It resolved the issue where the input and output order got mixed up
after compilation (a single line of code within `network_graph`).

To verify that this fix works as intended, I updated the
multi-input/output example and tested it, too.

If this change is applied, the current data-order problem will be fixed,
but it will affect the data order and weights order of apps that have
already been released.

therefore, do not merge this pr yet..
this commit just show one of the method how we can fix this issue.

Signed-off-by: SeungBaek Hong <sb92.hong@samsung.com>